Fix/back/delete (#62)
* chore: Add AssuntoRepository to DocumentoModule

* feat: Delete image when deleting Assunto or Documento

+++ echoeasy/src/repositories/documento.repository.ts
@@ -5,12 +5,14 @@ import { adminStorage } from 'src/config/firebase-admin';
 import { MulterFile } from 'src/types/File';
 import { DocumentoDto } from '../dto/DocumentoDto';
 import { Documento } from '../schema/Documento';
-
+import { AssuntoRepository } from './assunto.repository';
 @Injectable()
 export class DocumentoRepository {
   constructor(
     @InjectModel(Documento.name)
     private readonly documentoModel: Model<Documento>,
+
+    private readonly assuntoRepository: AssuntoRepository,
   ) {}
 
   async create(
@@ -148,6 +150,15 @@ export class DocumentoRepository {
       if (!Types.ObjectId.isValid(_id)) {
         throw new Error('ID inválido');
       }
+      const doc = await this.findOneById(_id);
+      if (!doc) {
+        throw new Error('Documento não encontrado');
+      }
+      const image_path = doc.image;
+      if (image_path) {
+        await this.deleteImage(image_path);
+      }
+      await this.assuntoRepository.deleteMannyByDocumentId(_id);
       return this.documentoModel.findOneAndDelete({ _id }).exec();
     } catch (error) {
       throw new HttpException(error.message, HttpStatus.NOT_FOUND);
@@ -215,6 +226,26 @@ export class DocumentoRepository {
 
         stream.end(file.buffer);
       });
+    } catch (error) {
+      throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
+    }
+  }
+
+  async deleteImage(imagePath: string): Promise<boolean> {
+    try {
+      if (!imagePath) {
+        throw new HttpException('Link Inválido', HttpStatus.BAD_REQUEST);
+      }
+
+      const parsedUrl = new URL(imagePath);
+      const relativePath = parsedUrl.pathname.replace(
+        '/echoeasy-539dc.appspot.com/',
+        '',
+      );
+
+      await adminStorage.file(relativePath).delete();
+
+      return true;
     } catch (error) {
       throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
     }
